refactor: create custom MDX Provider
