+++ src/components/mdx/index.ts
@@ -1,0 +1,1 @@
+export { default as MdxCustomProvider } from "./MdxCustomProvider";
